server/requirements.txt
Flask==3.0.3
pandas==2.2.2
pyarrow==16.1.0

python-dotenv==1.0.1
